:recycle: Move ImageContent component to component package
