+++ app/src/main/java/org/sopt/stamp/feature/mission/detail/MissionDetail.kt
@@ -1,21 +1,14 @@
 package org.sopt.stamp.feature.mission.detail
 
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.PickVisualMediaRequest
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.pager.HorizontalPager
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Button
 import androidx.compose.material.ButtonDefaults
@@ -27,13 +20,10 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.viewmodel.compose.viewModel
-import coil.compose.AsyncImage
 import com.airbnb.lottie.compose.LottieAnimation
 import com.airbnb.lottie.compose.LottieCompositionSpec
 import com.airbnb.lottie.compose.animateLottieCompositionAsState
@@ -50,86 +40,13 @@ import org.sopt.stamp.designsystem.component.toolbar.ToolbarIconType
 import org.sopt.stamp.designsystem.style.SoptTheme
 import org.sopt.stamp.domain.MissionLevel
 import org.sopt.stamp.feature.mission.detail.component.Header
+import org.sopt.stamp.feature.mission.detail.component.ImageContent
 import org.sopt.stamp.feature.mission.detail.component.Memo
 import org.sopt.stamp.feature.mission.model.ImageModel
 import org.sopt.stamp.feature.mission.model.MissionNavArgs
 import org.sopt.stamp.feature.ranking.getRankBackgroundColor
 import org.sopt.stamp.feature.ranking.getRankTextColor
 import org.sopt.stamp.util.DefaultPreview
-
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-private fun ImageContent(
-    imageModel: ImageModel,
-    onChangeImage: (images: ImageModel) -> Unit
-) {
-    val isImageEmpty = remember(imageModel) { imageModel.isEmpty() }
-    val photoPickerLauncher = rememberLauncherForActivityResult(ActivityResultContracts.PickMultipleVisualMedia()) {
-        onChangeImage(ImageModel.Local(it))
-    }
-    val pageCount = remember(imageModel) { imageModel.size }
-
-    HorizontalPager(pageCount) { page ->
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .aspectRatio(1f)
-                .background(
-                    color = SoptTheme.colors.onSurface5,
-                    shape = RoundedCornerShape(10.dp)
-                )
-        ) {
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .background(
-                        color = SoptTheme.colors.onSurface5,
-                        shape = RoundedCornerShape(10.dp)
-                    )
-                    .clickable {
-                        photoPickerLauncher.launch(
-                            PickVisualMediaRequest(ActivityResultContracts.PickVisualMedia.ImageOnly)
-                        )
-                    },
-                contentAlignment = Alignment.Center
-            ) {
-                if (isImageEmpty) {
-                    Text(
-                        text = "달성 사진을 올려주세요",
-                        style = SoptTheme.typography.sub2,
-                        color = SoptTheme.colors.onSurface50
-                    )
-                } else {
-                    when (imageModel) {
-                        is ImageModel.Local -> {
-                            AsyncImage(
-                                model = imageModel.uri[page],
-                                contentDescription = "",
-                                modifier = Modifier
-                                    .fillMaxSize()
-                                    .clip(RoundedCornerShape(10.dp)),
-                                contentScale = ContentScale.Crop
-                            )
-                        }
-
-                        is ImageModel.Remote -> {
-                            AsyncImage(
-                                model = imageModel.url[page],
-                                contentDescription = "",
-                                modifier = Modifier
-                                    .fillMaxSize()
-                                    .clip(RoundedCornerShape(10.dp)),
-                                contentScale = ContentScale.Crop
-                            )
-                        }
-
-                        else -> throw IllegalStateException("예외처리 했으므로 여긴 안 통과함")
-                    }
-                }
-            }
-        }
-    }
-}
 
 @MissionNavGraph
 @Destination("detail")
